+++ resources/js/bootstrap/stores/user.js
@@ -59,7 +59,6 @@ const updateUserAt = (context, userId, data, isProfile) => {
 
 export const useUserStore = defineStore('user', {
     state: _ => ({
-        userLoggedIn: false,
         user: {},
         users: [],
         deletedUsers: [],
@@ -69,6 +68,9 @@ export const useUserStore = defineStore('user', {
         preferences: {},
     }),
     getters: {
+        userLoggedIn: state => {
+            return !!state.user?.id;
+        },
         isSameUser: state => userId => {
             return state.user.id == userId;
         },
@@ -147,22 +149,17 @@ export const useUserStore = defineStore('user', {
                 return null;
             }
         },
-        setLoginState(value) {
-            this.userLoggedIn = value;
-        },
         setPreferences(preferences) {
             this.preferences = preferences;
         },
         async login(credentials) {
             await getCsrfCookie();
             const user = await login(credentials);
-            this.userLoggedIn = true;
             this.setActiveUser(user);
             await useSystemStore().initialize();
         },
         async logout() {
             await logout();
-            this.setLoginState(false);
             this.setActiveUser({});
         },
         setActiveUser(user, merge = false) {